Fix getJobName() string sanitization
This fixes the name sanitization using a regex.
The regex will substitute all the characters that are not
alphanumerics (`\\P{}`) with a "_".

This will avoid bugs when "Cache Dirs" volume names are generated.
For example, if a user has a Job inside a Folder the name will contain
the "Â" character. This is now substituted along all the previous
characters.