Refactoring and repackaging code
Renamed some files to more readable. Repackaging some classes to util or scheduler (for all objects based work on events or time) packages. Reimplemented way to generation name of the node. Now based on task label, global counter, and hashcode of the full name of the task converted to a radix-36 string.